support code blocks in statuses
